refactor(domains): move review, styleguide, and workflow modules into src
